doc: add section about regulatory.db missing
It was found that some distros do not ship with a regulatory.db file.
This is required to run some of the autotests, and if not found on the
system the kernel build will fail due to some of the IWD config
options which require regulatory.db.

The fix is to manually download regulatory.db from git.kernel.org and
place it in /lib/firmware.